Adding a prototype for Mentions input
The idea behihind this PR is to create a mentions input for prototyping in EH as well as the firefox fork. The firefox fork was able to impliment thier own place holder mentions with a 3rd party library. So this will continue to be a prototype for EH but I also intent to at least present this as an option to move into MC for our actual implimentation when the time comes. It has been created as a lit component with little deps and styleing so that it can be moved into MC with little effort.